feat: Add a Pulumi program to create the Github repository
This commit adds a configuration for the `tings` Github repository that
serves as the home of this project. In addition, the Dockerfile has been
updated to force Github to automatically associate the uploaded Docker
images with the repository. Finally, a Makefile has been added to
simplify the process of building and pushing a new container image to
Github.